refactor: reimplement OpenAI logprob streaming by replacing `openaiLogprobStream.ts` with a new `openaiLogprobs.ts` module and updating its usage
